Localisation heuristics (#30)
* Init heuristics for inferring system locale
* Revamp tests
